+++ biome.json
@@ -58,8 +58,7 @@
     "enabled": true,
     "lineWidth": 80,
     "indentWidth": 2,
-    "indentStyle": "space",
-    "lineEnding": "crlf"
+    "indentStyle": "space"
   },
   "files": {
     "ignore": [".next/*", "main.ts"],
