+++ src/view/login/component/LoginForm/index.css.ts
@@ -1,4 +1,3 @@
-import { theme } from "@/styles/themes.css";
 import { style } from "@vanilla-extract/css";
 
 export const formStyle = style({
@@ -20,10 +19,4 @@ export const contentStyle = style({
 export const itemStyle = style({
   height: "8rem",
   justifyContent: "flex-end",
-});
-
-export const fieldStyle = style({
-  "::placeholder": {
-    color: theme.color.white,
-  },
 });
